good version

+++ MINI SITE/deleteMsg.aspx.cs
@@ -19,34 +19,41 @@ namespace MINI_SITE
 
         protected void Page_Load(object sender, EventArgs e)
         {
-            string id = Request.QueryString["Id"];
-            if (id != null && id != "")
+            if (Session["login_user"] != null)
             {
-                if (Request["submitbtn"] == null)
+                string id = Request.QueryString["Id"];
+                if (id != null && id != "")
                 {
-                    DataTable dt = SQLHelper.SelectData(string.Format("select subject, message, sender from inbox where id = N'{0}'", id));
-                    DataRow row;
-                    if (dt.Rows.Count >= 1)
+                    if (Request["submitbtn"] == null)
                     {
-                        row = dt.Rows[0];
-                        subject = row["subject"].ToString();
-                        message = row["message"].ToString();
-                        sender1 = row["sender"].ToString();                       
+                        DataTable dt = SQLHelper.SelectData(string.Format("select subject, message, sender from inbox where id = N'{0}'", id));
+                        DataRow row;
+                        if (dt.Rows.Count >= 1)
+                        {
+                            row = dt.Rows[0];
+                            subject = row["subject"].ToString();
+                            message = row["message"].ToString();
+                            sender1 = row["sender"].ToString();
+                        }
+                    }
+                    else
+                    {
+                        string sql = "delete from inbox where id = N'" + id + "'";
+                        SQLHelper.DoQuery(sql);
+                        Response.Redirect("deleteMsg.aspx?message=message " + subject + " was deleted, redirecting to inbox page");
                     }
                 }
                 else
                 {
-                    string sql = "delete from inbox where id = N'" + id + "'";
-                    SQLHelper.DoQuery(sql);
-                    Response.Redirect("deleteMsg.aspx?message=message " + subject + " was deleted, redirecting to inbox page");
+                    msg = Request.QueryString["message"];
+                    makeTableInvisible = "style='display : none'";
+                    redirectJs = "setTimeout(\"location.href ='inbox.aspx';\", 3000);";
+
                 }
             }
             else
             {
-                msg = Request.QueryString["message"];
-                makeTableInvisible = "style='display : none'";
-                redirectJs = "setTimeout(\"location.href ='inbox.aspx';\", 3000);";
-
+                Response.Redirect("login.aspx");
             }
         }
     }
